requirements.txt
ultralytics>=8.4.41
torch>=2.0.0
torchvision>=0.15.0
Pillow>=10.0.0
PyYAML>=6.0.0
opencv-python>=4.8.0
paddleocr>=2.8.1
streamlit>=1.30.0
numpy>=1.24.0
